go/oasis-node/cmd/storage: Factor compact comand to separate file
